+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 edbee.lib:
 
+- refs #19 Workaround-hack for non-bmp-unicode character movement. (experimental)
+- Added DebugCommand::DumpCharacterCodes (for dumping hex-character codes)
 - fix #14, MinGW compatibility: Disabled memoryLeak detection and fixed mingw compilation (different library-name is generated for mingw)
 - fix #13, Added a method to disable the scrollarea shadows: widget->textScrollArea()->enableShadowWidget(bool)
 - fix #9, Updated onigmo library. Fixes compilation/linkage issue on Mac OS X (enc/windows_31j.c)